removed extra LetterResult file
