+++ wordle-web/components/GameboardGuess.vue
@@ -1,5 +1,5 @@
 <template>
-  <LetterResult
+  <GameboardLetter
     v-for="(letter, j) of guess.letters"
     :key="j"
     :letter="letter"
@@ -9,6 +9,7 @@
 
 <script setup lang="ts">
 import { defineProps } from "vue";
+import Game from "~/pages/game.vue";
 import { Word } from "~/scripts/word";
 defineProps<{
   guess: Word;
